+++ api/v1/views/__init__.py
@@ -15,4 +15,4 @@ from api.v1.views.cities import *
 from api.v1.views.amenities import *
 from api.v1.views.places import *
 from api.v1.views.users import *
-from api.v1.views.reviews import *
+from api.v1.views.places_reviews import *
