--- api/v1/views/reviews.py
@@ -1,93 +1,0 @@
-#!/usr/bin/python3
-
-
-"""
-This module provides RESTful API actions for Review objects.
-It includes routes to retrieve, create, delete, and update reviews for places.
-"""
-
-from flask import jsonify, abort, request
-from api.v1.views import app_views
-from models import storage
-from models.place import Place
-from models.review import Review
-from models.user import User
-
-
-@app_views.route('/places/<place_id>/reviews', methods=['GET'])
-def get_reviews_by_place(place_id):
-    """Retrieves the list of all Review objects of a Place."""
-    place = storage.get(Place, place_id)
-    if not place:
-        abort(404)
-    reviews = [review.to_dict() for review in place.reviews]
-    return jsonify(reviews)
-
-
-@app_views.route('/reviews/<review_id>', methods=['GET'])
-def get_review(review_id):
-    """Retrieves a specific Review by ID."""
-    review = storage.get(Review, review_id)
-    if not review:
-        abort(404)
-    return jsonify(review.to_dict())
-
-
-@app_views.route('/reviews/<review_id>', methods=['DELETE'])
-def delete_review(review_id):
-    """Deletes a Review by ID."""
-    review = storage.get(Review, review_id)
-    if not review:
-        abort(404)
-    storage.delete(review)
-    storage.save()
-    return jsonify({}), 200
-
-
-@app_views.route('/places/<place_id>/reviews', methods=['POST'])
-def create_review(place_id):
-    """Creates a new Review under a specific Place."""
-    place = storage.get(Place, place_id)
-    if not place:
-        abort(404)
-
-    if not request.is_json:
-        abort(400, description="Not a JSON")
-
-    data = request.get_json()
-    if 'user_id' not in data:
-        abort(400, description="Missing user_id")
-    if 'text' not in data:
-        abort(400, description="Missing text")
-
-    user = storage.get(User, data['user_id'])
-    if not user:
-        abort(404)
-
-    new_review = Review(text=data['text'], place_id=place_id, user_id=data['user_id'])
-    for key, value in data.items():
-        if key not in ['id', 'user_id', 'place_id', 'created_at', 'updated_at']:
-            setattr(new_review, key, value)
-
-    storage.new(new_review)
-    storage.save()
-    return jsonify(new_review.to_dict()), 201
-
-
-@app_views.route('/reviews/<review_id>', methods=['PUT'])
-def update_review(review_id):
-    """Updates a Review by ID."""
-    review = storage.get(Review, review_id)
-    if not review:
-        abort(404)
-
-    if not request.is_json:
-        abort(400, description="Not a JSON")
-
-    data = request.get_json()
-    for key, value in data.items():
-        if key not in ['id', 'user_id', 'place_id', 'created_at', 'updated_at']:
-            setattr(review, key, value)
-
-    storage.save()
-    return jsonify(review.to_dict()), 200
